% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{pairs.stanidm}
\alias{pairs.stanidm}
\title{Pairs method for stanidm objects}
\usage{
\method{pairs}{stanidm}(x, pars = NULL, regex_pars = NULL,
  condition = pairs_condition(nuts = "accept_stat__"), ...)
}
\description{
Interface to \pkg{bayesplot}'s \code{\link[bayesplot]{mcmc_pairs}} function
for use with \pkg{rstanarm} models. Be careful not to specify too
many parameters to include or the plot will be both hard to read and slow to
render.
}
