% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_survfit.R
\name{plot.fit.stanidm}
\alias{plot.fit.stanidm}
\title{Plot method for stanidm
 -------------------------------------------}
\usage{
\method{plot}{survfit.stansurv}(object, ...)
}
\description{
Plot the estimated subject-specific or marginal
 survival function for each transition hazard
}
