% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sclc_demo}
\alias{sclc_demo}
\title{SCLC
A dataset containing the survival outcome in SCLC trial}
\format{Survival outcome data frame (time, cens).
\describe{
  \item{time}{Actual survival months observed}
  \item{cens}{censoring indicator 1 observed, 0 censored}
  ...
}}
\usage{
sclc_demo
}
\description{
SCLC
A dataset containing the survival outcome in SCLC trial
}
\keyword{datasets}
